Use shared rippledSource module in compareHostFunctions.js
